Allow building with older Docker versions
Older Docker versions (as available in many distributions) require
their descriptor files to be inside their work directory. This is the
only impediment to using these versions of Docker with the OLM local
build infrastructure, so this patch copies the descriptors into the
work directory (bin, which is deleted on clean-up) and references the
copies.

Signed-off-by: Stephen Kitt <skitt@redhat.com>